requirements2.txt
agno==1.2.4
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.49.0
anyascii==0.3.2
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
boto3==1.36.2
botocore==1.36.2
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.0
celery==5.4.0
celery-progress==0.5
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
daphne==4.1.2
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
dj-rest-auth==7.0.1
dj-stripe==2.8.4
Django==5.1.5
django-allauth==65.3.1
django-anymail==12.0
django-appconf==1.1.0
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-cryptography==1.1
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-filter==24.3
django-health-check==3.18.3
django-hijack==3.7.1
django-modelcluster==6.4
django-permissionedforms==0.1
django-storages==1.14.4
django-taggit==6.1.0
django-timezone-field==7.1
django-treebeard==4.7.1
django-waffle==4.2.0
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
djangorestframework_simplejwt==5.4.0
docstring_parser==0.16
docutils==0.21.2
draftjs-exporter==5.0.0
drf-spectacular==0.28.0
duckduckgo_search==2025.4.4
et_xmlfile==2.0.0
fido2==1.2.0
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-genai==1.9.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.22.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
identify==2.6.6
idna==3.10
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jira==3.8.0
jiter==0.8.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
l18n==2021.3
laces==0.1.2
lxml==5.3.1
mailchimp3==3.0.21
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
nodeenv==1.9.1
numpy==2.2.4
oauthlib==3.2.2
openai==1.69.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pgvector==0.4.0
pillow==11.1.0
pillow_heif==0.21.0
pip-tools==7.4.1
platformdirs==4.3.6
pre_commit==4.1.0
primp==0.14.0
prompt_toolkit==3.0.50
proto-plus==1.25.0
protobuf==5.29.3
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyproject_hooks==1.2.0
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
redis==5.2.1
referencing==0.36.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.9.7
s3transfer==0.11.1
selenium==4.31.0
sentry-sdk==2.20.0
service-identity==24.2.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
slack_sdk==3.35.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
stripe==4.2.0
telepath==0.3.1
tomli==2.2.1
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
txaio==23.1.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
virtualenv==20.29.1
wagtail==6.3.2
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.8.2
Willow==1.9.0
wsproto==1.2.0
zope.interface==7.2
